Partially back out b6612b6
The DELETE action on the /alerts endpoint turns out to be unnecessary.
Zabbix's alert framework doesn't make a distinction between a new alert
and a resolved alert so there's no way for a consumer of the Zpark API
to actually choose which verb to use.

For now, everything will use POST.