Standardize on camelCase for JSON output
An upcoming commit will add more items to the /alert endpoint's output
making it necessary to decide on camelCase or under_case formats. Camel
case keeps things consistent with the Cisco Spark API.

This is technically a modification to the Zpark API, but since this is
pre-release code, I'm not cranking the version number.