Push handling of the Spark API objects down to the task layer
This enables the Zpark API layer to stay simple and generic; it simply
passes around general data such as message contents and who the message
should be addressed to and the task layer takes care of packaging that
information into something that the Spark API module can deal with. This
will be more important as work continues on the support for Spark's
webhooks.