Valid incoming webhook callbacks have a good X-Spark-Signature
With this commit, the application will only process webhook callbacks
from the Spark service by ensuring the callback has set the
X-Spark-Signature header and the header value equals the correct SHA1
HMAC of the request body

This check can be disabled by unsetting the 'SPARK_WEBHOOK_SECRET'
setting in app.cfg. When this feature is *enabled*, make sure you're
creating your webhooks with the 'secret' parameter set to the same value
as 'SPARK_WEBHOOK_SECRET'.

+tests