Fix passing of objects across the app/task boundary
The default serializer in Celery is a JSON-based one. It does not
support serialization of objects which directly conflicts with what I'm
trying to do. Unfortunately, unit tests didn't find this because I don't
test the tasks in async mode.

Fix this by converting the Spark resource objects (rooms, etc) to a dict
and passing the dict around.