Test coverage for a specific api->task codepath
This catches some fallout related to the passing around of room and
people objects: send_spark_alert_message() in the API layer was not
passing a proper room/person construct to the task layer which was
causing an uncaught exception to be thrown. This broke all alerts coming
in from zabbix.

This commit adds 2 test cases which catch this behavior.