Define logging config via logging dict and switch to syslog
This commit switches all log handlers from rotating file to syslog
(facility local6). The logging config is now defined in a dict which is
loaded using logging.config.dictConfig(). This gives more configuration
power to the user and means the syntax is already documented courtesy of
the logging.config module docs :) This commit also changes the
formatting of log messages slightly.