ZPARK_API_TOKEN can no longer be disabled
This was a poor design choice in 7682a27 that allowed the token to be
disabled. There is no good use case for that and is a highly risky
configuration. This commit changes the semantics so that the token will
reject all API requests when set to None. This also becomes the sane
default value.

While here, tweak some log messages for clarity.