Allow [;:,] as delimiters between @bot and the command
Part of this commit also refactors some of the validation of received
commands to come after some of the processing (determing bot name and
accounting or any delimiters). This way, validation can happen on the
exact command and not (potentially, when having received a message in a
group room) on the '@bot, ' beginning part of the message.

Unfortunately to do things in this order we do have to query the Spark
API for the room information prior to command validation which means the
query is pure overhead if we throw the command away. Leave the query for
the message info until after validation for this very reason.

For #5

With tests